Incompatible change! Move config info from version file to new config file.
The version number and chunk size were moved from the version files
into a new config file because they should really not be changed
after the initial "init" of the backup. Also changed the enc config file's
magic string.

This version is incompatible with backups done with older versions.
This is acceptable as this is still alpha software.